change converters to use multiple binding and pass the style to verify if the console is in the dark mode configuration
