+++ src/Lime.Client.TestConsole/Converters/IsRawToBrushConverter.cs
@@ -4,29 +4,35 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Data;
 using System.Windows.Media;
 
 namespace Lime.Client.TestConsole.Converters
 {
-    public class IsRawToBrushConverter : IValueConverter
+    public class IsRawToBrushConverter : IMultiValueConverter
     {
         #region IValueConverter Members
 
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
+        public object Convert(object[] values, Type targetType, object parameter, System.Globalization.CultureInfo culture)
         {
-            if (value is bool)
+            //First value is IsRaw property
+            //Second value is IsDarkMode property
+
+            if (values[0] is bool && (bool)values[0])
             {
-                if ((bool)value)
-                {
-                    return new SolidColorBrush(Colors.Red);
-                }                                
+                return new SolidColorBrush(Colors.Red);
+            }
+
+            if (values[1] is Style)
+            {
+                return new SolidColorBrush(Colors.White);
             }
 
             return new SolidColorBrush(Colors.Black);
         }
 
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
+        public object[] ConvertBack(object value, Type[] targetTypes, object parameter, System.Globalization.CultureInfo culture)
         {
             throw new NotImplementedException();
         }
